[redis] Test are guaranteed to run on all configs
I made it so all queries end up in the slowlog during the slowlog test,
putting 0s as the slowlog logging threshold. This ensures that these
tests succeed, even on a very very fast Redis configuration. The slowlog
logging threshold is reset to its original value at the end of the
tests.